+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Enhance ScanPayload interface by adding barcode property and updating pass type to allow null values for non-a2w barcodes.. (`b63f4b6`)
+
 ## 0.2.4 - 2026-05-15
 
 - Update repository URL in package.json and adjust release workflow comments to clarify provenance verification requirements.. (`9beef27`)